\name{cplxrv-package}
\alias{cplxrv-package}
\alias{cplxrv}
\docType{package}
\title{
\packageTitle{cplxrv}
}
\description{
\packageDescription{cplxrv}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{cplxrv}

Maintainer: \packageMaintainer{cplxrv}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = cplxrv)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{cplxrv}}
%\section{Documentation Index}{\packageIndices{cplxrv}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
