% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcnorm.R
\name{dcnorm}
\alias{dcnorm}
\title{Complex-valued Density Representation}
\usage{
dcnorm(z, mean, var, rel, is.circular = FALSE, log = FALSE)
}
\arguments{
\item{z}{Complex-valued input (vector or matrix)}

\item{mean}{Mean of the distribution}

\item{var}{Variance-covariance matrix}

\item{rel}{Relation matrix (only used when is.circular = FALSE)}

\item{is.circular}{Logical; if TRUE, uses circular complex normal distribution}

\item{log}{Logical; if TRUE, returns log-density}
}
\value{
Density or log-density value(s)
}
\description{
Complex-valued Density Representation
}
